restructured files
moved add-workkouts-to-cycle code into proper place in cycle folder
